chore: Move php app helper code to separate file
